requirements.txt
absl-py==0.8.1
albumentations==0.4.3
backcall==0.1.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
google-auth==1.8.2
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
imageio==2.6.1
imgaug==0.2.6
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
prompt-toolkit==3.0.2
protobuf==3.11.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.3.3
six==1.13.0
tabulate==0.8.6
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
